# 8285633: Take better advantage of generic MethodType cache