--- a/src/java.base/share/classes/java/lang/invoke/MethodTypeForm.java
+++ b/src/java.base/share/classes/java/lang/invoke/MethodTypeForm.java
@@ -194,7 +194,7 @@ final class MethodTypeForm {
             this.lambdaForms   = new SoftReference[LF_LIMIT];
             this.methodHandles = new SoftReference[MH_LIMIT];
         } else {
-            this.basicType = MethodType.makeImpl(basicReturnType, basicPtypes, true);
+            this.basicType = MethodType.methodType(basicReturnType, basicPtypes, true);
             // fill in rest of data from the basic type:
             MethodTypeForm that = this.basicType.form();
             assert(this != that);
@@ -250,7 +250,7 @@ final class MethodTypeForm {
         // Find the erased version of the method type:
         if (rtypeCanonical == null)  rtypeCanonical = rtype;
         if (ptypesCanonical == null)  ptypesCanonical = ptypes;
-        return MethodType.makeImpl(rtypeCanonical, ptypesCanonical, true);
+        return MethodType.methodType(rtypeCanonical, ptypesCanonical, true);
     }
 
     /** Canonicalize the given return or param type.